edited dashboard need to do internationalization
