+++ shift-scan/src/app/(routes)/dashboard/UI/_buttons/AdditonalclockOutBtns.tsx
@@ -38,156 +38,161 @@ export default function ClockOutWidget({
   const c = useTranslations("Clock");
   return (
     <>
-      <Holds className="col-span-2 row-span-1 gap-5 h-full">
-        <Buttons background={"orange"} onClick={() => setIsModal2Open(true)}>
-          <Holds position={"row"} className="my-auto">
-            <Holds size={"60"}>
-              <Texts size={"p1"}>{t("Break")}</Texts>
+      <Grids rows={"3"} gap={"5"}>
+        <Holds className="col-span-2 row-span-1 gap-5 h-full">
+          <Buttons background={"orange"} onClick={() => setIsModal2Open(true)}>
+            <Holds position={"row"} className="my-auto">
+              <Holds size={"60"}>
+                <Texts size={"p1"}>{t("Break")}</Texts>
+              </Holds>
+              <Holds size={"40"}>
+                <Images
+                  titleImg="/break.svg"
+                  titleImgAlt="Break Icon"
+                  size={"50"}
+                />
+              </Holds>
             </Holds>
-            <Holds size={"40"}>
-              <Images
-                titleImg="/break.svg"
-                titleImgAlt="Break Icon"
-                size={"50"}
-              />
-            </Holds>
-          </Holds>
-        </Buttons>
-      </Holds>
-      <NModals
-        isOpen={isModal2Open}
-        handleClose={() => setIsModal2Open(false)}
-        size={"screen"}
-        background={"takeABreak"}
-      >
-        <Holds background={"white"} className="h-full w-full">
-          <Holds background={"white"} className="h-full w-[90%] p-1">
-            <Grids rows={"7"} gap={"5"} className="mb-5 h-full w-full">
-              <Holds className="row-start-1 row-end-2 h-full w-full justify-center">
-                <Grids
-                  rows={"1"}
-                  cols={"5"}
-                  gap={"3"}
-                  className=" h-full w-full"
-                >
-                  <Holds
-                    className="row-start-1 row-end-2 col-start-1 col-end-2 h-full w-full justify-center"
-                    onClick={() => setIsModal2Open(false)}
-                  >
-                    <Images
-                      titleImg="/turnBack.svg"
-                      titleImgAlt="back"
-                      position={"left"}
-                    />
-                  </Holds>
-                </Grids>
-              </Holds>
-
-              <Holds className="row-start-2 row-end-4 h-full w-full justify-center relative">
-                <Holds className="h-full w-[90%] relative">
-                  <Labels size={"p4"} htmlFor="comment">
-                    {c("PreviousJobComment")}
-                  </Labels>
-                  <TextAreas
-                    value={comment}
-                    onChange={(e) => setComment(e.target.value)}
-                    placeholder={c("TodayIDidTheFollowing")}
-                    className="w-full h-full"
-                    maxLength={40}
-                    style={{ resize: "none" }}
-                  />
-
-                  <Texts
-                    size={"p2"}
-                    className={`${
-                      comment.length === 40
-                        ? "text-red-500 absolute bottom-5 right-2"
-                        : "absolute bottom-5 right-2"
-                    }`}
-                  >
-                    {comment.length}/40
-                  </Texts>
-                </Holds>
-              </Holds>
-
-              <Holds position={"row"} className="row-start-7 row-end-8 h-full ">
-                <Buttons
-                  background={"orange"}
-                  onClick={() => handleCOButton2()}
-                  className="w-full py-4 "
-                >
-                  <Titles size={"h2"}>{c("Continue")}</Titles>
-                </Buttons>
-              </Holds>
-            </Grids>
-          </Holds>
+          </Buttons>
         </Holds>
-      </NModals>
-
-      <Holds className="col-span-2 row-span-1 gap-5 h-full">
-        <Buttons background={"red"} onClick={handleCOButton3}>
-          <Holds position={"row"} className="my-auto">
-            <Holds size={"70"}>
-              <Texts size={"p1"}>{t("EndDay")}</Texts>
-            </Holds>
-            <Holds size={"30"}>
-              <Images
-                titleImg="/end-day.svg"
-                titleImgAlt="End Icon"
-                size={"50"}
-              />
-            </Holds>
-          </Holds>
-        </Buttons>
-      </Holds>
-      <Modals
-        isOpen={isModalOpen}
-        handleClose={handleCloseModal}
-        size={"clock"}
-      >
-        <Bases>
-          <Contents>
-            <Holds background={"white"} className="h-full">
-              <Holds className="h-full py-10">
-                <Contents width={"section"}>
-                  <Grids rows={"4"} gap={"5"}>
-                    <Holds className="h-full span-3 my-auto">
-                      <Titles size={"h1"}>{t("Whoops")}</Titles>
-                      <br />
-                      <Texts size={"p2"}>{t("ReturnToLogOut")}</Texts>
-                    </Holds>
-                    <Holds className="h-full span-1 my-auto">
-                      <Buttons
-                        background={"orange"}
-                        size={"full"}
-                        href={`/dashboard/equipment`}
-                      >
-                        <Texts size={"p3"}>{t("ClickToLogOut")}</Texts>
-                      </Buttons>
+        <NModals
+          isOpen={isModal2Open}
+          handleClose={() => setIsModal2Open(false)}
+          size={"screen"}
+          background={"takeABreak"}
+        >
+          <Holds background={"white"} className="h-full w-full">
+            <Holds background={"white"} className="h-full w-[90%] p-1">
+              <Grids rows={"7"} gap={"5"} className="mb-5 h-full w-full">
+                <Holds className="row-start-1 row-end-2 h-full w-full justify-center">
+                  <Grids
+                    rows={"1"}
+                    cols={"5"}
+                    gap={"3"}
+                    className=" h-full w-full"
+                  >
+                    <Holds
+                      className="row-start-1 row-end-2 col-start-1 col-end-2 h-full w-full justify-center"
+                      onClick={() => setIsModal2Open(false)}
+                    >
+                      <Images
+                        titleImg="/turnBack.svg"
+                        titleImgAlt="back"
+                        position={"left"}
+                      />
                     </Holds>
                   </Grids>
-                </Contents>
-              </Holds>
-            </Holds>
-          </Contents>
-        </Bases>
-      </Modals>
-      <Holds className="col-span-2 row-span-1 gap-5 h-full">
-        <Buttons background={"lightBlue"} onClick={handleShowManagerButtons}>
-          <Holds position={"row"} className="my-auto">
-            <Holds size={"60"}>
-              <Texts size={"p1"}>{t("GoHome")}</Texts>
-            </Holds>
-            <Holds size={"40"}>
-              <Images
-                titleImg="/home.svg"
-                titleImgAlt="Home Icon"
-                size={"50"}
-              />
+                </Holds>
+
+                <Holds className="row-start-2 row-end-4 h-full w-full justify-center relative">
+                  <Holds className="h-full w-[90%] relative">
+                    <Labels size={"p4"} htmlFor="comment">
+                      {c("PreviousJobComment")}
+                    </Labels>
+                    <TextAreas
+                      value={comment}
+                      onChange={(e) => setComment(e.target.value)}
+                      placeholder={c("TodayIDidTheFollowing")}
+                      className="w-full h-full"
+                      maxLength={40}
+                      style={{ resize: "none" }}
+                    />
+
+                    <Texts
+                      size={"p2"}
+                      className={`${
+                        comment.length === 40
+                          ? "text-red-500 absolute bottom-5 right-2"
+                          : "absolute bottom-5 right-2"
+                      }`}
+                    >
+                      {comment.length}/40
+                    </Texts>
+                  </Holds>
+                </Holds>
+
+                <Holds
+                  position={"row"}
+                  className="row-start-7 row-end-8 h-full "
+                >
+                  <Buttons
+                    background={"orange"}
+                    onClick={() => handleCOButton2()}
+                    className="w-full py-4 "
+                  >
+                    <Titles size={"h2"}>{c("Continue")}</Titles>
+                  </Buttons>
+                </Holds>
+              </Grids>
             </Holds>
           </Holds>
-        </Buttons>
-      </Holds>
+        </NModals>
+
+        <Holds className="col-span-2 row-span-1 gap-5 h-full">
+          <Buttons background={"red"} onClick={handleCOButton3}>
+            <Holds position={"row"} className="my-auto">
+              <Holds size={"70"}>
+                <Texts size={"p1"}>{t("EndDay")}</Texts>
+              </Holds>
+              <Holds size={"30"}>
+                <Images
+                  titleImg="/end-day.svg"
+                  titleImgAlt="End Icon"
+                  size={"50"}
+                />
+              </Holds>
+            </Holds>
+          </Buttons>
+        </Holds>
+        <Modals
+          isOpen={isModalOpen}
+          handleClose={handleCloseModal}
+          size={"clock"}
+        >
+          <Bases>
+            <Contents>
+              <Holds background={"white"} className="h-full">
+                <Holds className="h-full py-10">
+                  <Contents width={"section"}>
+                    <Grids rows={"4"} gap={"5"}>
+                      <Holds className="h-full span-3 my-auto">
+                        <Titles size={"h1"}>{t("Whoops")}</Titles>
+                        <br />
+                        <Texts size={"p2"}>{t("ReturnToLogOut")}</Texts>
+                      </Holds>
+                      <Holds className="h-full span-1 my-auto">
+                        <Buttons
+                          background={"orange"}
+                          size={"full"}
+                          href={`/dashboard/equipment`}
+                        >
+                          <Texts size={"p3"}>{t("ClickToLogOut")}</Texts>
+                        </Buttons>
+                      </Holds>
+                    </Grids>
+                  </Contents>
+                </Holds>
+              </Holds>
+            </Contents>
+          </Bases>
+        </Modals>
+        <Holds className="col-span-2 row-span-1 gap-5 h-full">
+          <Buttons background={"lightBlue"} onClick={handleShowManagerButtons}>
+            <Holds position={"row"} className="my-auto">
+              <Holds size={"60"}>
+                <Texts size={"p1"}>{t("GoHome")}</Texts>
+              </Holds>
+              <Holds size={"40"}>
+                <Images
+                  titleImg="/home.svg"
+                  titleImgAlt="Home Icon"
+                  size={"50"}
+                />
+              </Holds>
+            </Holds>
+          </Buttons>
+        </Holds>
+      </Grids>
     </>
   );
 }
