+++ shift-scan/src/app/(routes)/dashboard/UI/_buttons/clockOutBtn.tsx
@@ -21,7 +21,7 @@ export default function ClockOutBtn({
         permission !== "USER" && View === "general"
           ? "row-span-1 col-span-1 gap-5"
           : permission !== "USER" && View === "mechanic"
-          ? "row-span-1 col-span-2 gap-5"
+          ? "row-span-1 col-span-1 gap-5"
           : permission !== "USER" && View === "truck"
           ? "row-span-1 col-span-1 gap-5"
           : permission !== "USER" && View === "equipment"
@@ -43,6 +43,8 @@ export default function ClockOutBtn({
               : permission !== "USER" && View === "truck"
               ? undefined
               : permission === "USER" && View === "truck"
+              ? undefined
+              : permission !== "USER" && View === "mechanic"
               ? undefined
               : "row"
           }
