+++ shift-scan/src/app/(routes)/dashboard/UI/_dashboards/mechanicDashboardView.tsx
@@ -12,7 +12,8 @@ import GeneratorBtn from "../_buttons/generatorBtn";
 import MyTeamWidget from "../_buttons/myTeamBtn";
 import SwitchJobsBtn from "../_buttons/switchJobsBtn";
 import { Dispatch, SetStateAction } from "react";
-import EngineerBtn from "../_buttons/EnginneerBtns";
+import EngineerBtn from "../_buttons/MechanicBtns";
+import MechanicBtn from "../_buttons/MechanicBtns";
 
 export default function MechanicDashboardView({
   additionalButtonsType,
@@ -44,27 +45,44 @@ export default function MechanicDashboardView({
   return (
     <>
       <Contents width={"section"} className="py-5">
-        <Grids cols={"2"} rows={permission !== "USER" ? "4" : "3"} gap={"5"}>
+        <Grids cols={"2"} rows={"3"} gap={"5"}>
           {/* Render buttons based on state */}
           {additionalButtonsType === "equipment" ? (
-            <EquipmentWidget
-              handleShowManagerButtons={handleShowManagerButtons}
-            />
+            <Holds
+              className={
+                permission !== "USER"
+                  ? "col-span-2 row-span-4 gap-5 h-full"
+                  : "col-span-2 row-span-3 gap-5 h-full"
+              }
+            >
+              <EquipmentWidget
+                handleShowManagerButtons={handleShowManagerButtons}
+              />
+            </Holds>
           ) : additionalButtonsType === "clockOut" ? (
-            <ClockOutWidget
-              handleShowManagerButtons={handleShowManagerButtons}
-              setIsModal2Open={setIsModal2Open}
-              isModal2Open={isModal2Open}
-              isModalOpen={isModalOpen}
-              comment={comment}
-              setComment={setComment}
-              handleCOButton2={handleCOButton2}
-              handleCOButton3={handleCOButton3}
-              handleCloseModal={handleCloseModal}
-            />
+            <Holds
+              className={
+                permission !== "USER"
+                  ? "col-span-2 row-span-4 gap-5 h-full"
+                  : "col-span-2 row-span-3 gap-5 h-full"
+              }
+            >
+              <ClockOutWidget
+                handleShowManagerButtons={handleShowManagerButtons}
+                setIsModal2Open={setIsModal2Open}
+                isModal2Open={isModal2Open}
+                isModalOpen={isModalOpen}
+                comment={comment}
+                setComment={setComment}
+                handleCOButton2={handleCOButton2}
+                handleCOButton3={handleCOButton3}
+                handleCloseModal={handleCloseModal}
+              />
+            </Holds>
           ) : (
             <>
-              <EngineerBtn permission={permission} view={"mechanic"} />
+              <MechanicBtn permission={permission} view={"mechanic"} />
+              <SwitchJobsBtn permission={permission} />
               {permission !== "USER" && !additionalButtonsType && (
                 <GeneratorBtn />
               )}
@@ -73,8 +91,6 @@ export default function MechanicDashboardView({
               )}
 
               <FormsBtn permission={permission} view={"mechanic"} />
-
-              <SwitchJobsBtn permission={permission} />
 
               <ClockOutBtn
                 handleShowAdditionalButtons={handleShowAdditionalButtons}
